+++ lib/health_web/templates/exercise/index.html.eex
@@ -8,23 +8,23 @@
   <hr>
 <% end %>
 
-<%= HealthWeb.LayoutView.tabs fn f -> %>
-  <%= HealthWeb.LayoutView.tab(f, "One") do %>
+<%= tabs fn f -> %>
+  <%= tab(f, "One") do %>
     <p>This is tab one</p>
   <% end %>
 
-  <%= HealthWeb.LayoutView.tab(f, "Two") do %>
+  <%= tab(f, "Two") do %>
     <p>This is tab two</p>
   <% end %>
 <% end %>
 
 
-<%= HealthWeb.LayoutView.tabs [activate: "Two", class: "coolman"], fn f -> %>
-  <%= HealthWeb.LayoutView.tab(f, "One") do %>
+<%= tabs [activate: "Two", class: "coolman"], fn f -> %>
+  <%= tab(f, "One") do %>
     <p>This is tab one</p>
   <% end %>
 
-  <%= HealthWeb.LayoutView.tab(f, "Two") do %>
+  <%= tab(f, "Two") do %>
     <p>This is tab two</p>
   <% end %>
 <% end %>
